+++ cli/cli/command_str_consts/command_str_consts.go
@@ -39,5 +39,6 @@ const (
 		InitCmdStr = "init"
 		PathCmdStr = "path"
 	VersionCmdStr = "version"
+	GatewayCmdStr = "gateway"
 )